control flow transformations
For the two if statements, note that both branches execute the same statements
before and after. For the while loop, note that it is an infinite loop
and there are no internal `continue`s so it is safe to roll `q = p` from
the beginning to the end and before the loop.