bugfix: memory safety after bug()
It is an easily-forgotten fact that bug() is not a `noreturn` function,
because the user can choose to continue after hitting a bug. Therefore,
code must be careful not to trigger undefined behavior after a `bug()`
check, and should always do their best to exit the current scope
without undue damage and ideally without leaking memory either.