refactor: getFullArg (#56)
* make getFullArg private

* move comment

* move free of temps to just after last use

To check this, ensure that any code path ending in `goto return0`
or `goto return1` either frees `tmpStr` and `tmpArg` or never sets them.

* unify two return paths

To verify, note that each `goto return0` becomes
`ret = 0; goto getFullArg_ret`  and `goto return1` becomes
` goto getFullArg_ret`, because the initial value of `ret` is 1.

* control flow transformations

For the two if statements, note that both branches execute the same statements
before and after. For the while loop, note that it is an infinite loop
and there are no internal `continue`s so it is safe to roll `q = p` from
the beginning to the end and before the loop.

* move down temporary declarations

To verify: observe that it still compiles, nothing is out of scope.

* move integer declarations down

Note that `j` was split to another variable `j2` at the end of the
function. To verify, note that all subsequent uses of `j` have become
`j2`.

* move file down

* improve variable names

This was done by 'rename variable' tool. (Note that we couldn't have
done this before splitting temporaries used for multple purposes.)

* document why this isn't a memory leak

The code was doing something tricky here that should be called out
explicitly. To verify: `defaultCmd` is set inside the
`if (cmdList[0] == '#')` and
 `if (cmdList[0] == '*' || cmdList[0] == '&')`
 conditionals, but in both cases there is an unconditional jump at the
 end of the if, so any path that makes it to the commented
 line has still not set `defaultCmd`.

* bugfix: memory safety after bug()

It is an easily-forgotten fact that bug() is not a `noreturn` function,
because the user can choose to continue after hitting a bug. Therefore,
code must be careful not to trigger undefined behavior after a `bug()`
check, and should always do their best to exit the current scope
without undue damage and ideally without leaking memory either.

* use else if for mutually exclusive options

* drop tmpArg