Simplify and harden `switchPos`
The original implementation concatenated all the arguments with spaces,
which is not correct (causes a segfault) when the arguments themselves
can have spaces. Here we use a much simpler mechanism, just searching
for `/` and then `MY_SWITCH`, taking advantage of the fact that every
existing switch is one word.

fixes #136